Rewrite the table parsing to more closely follow cmark-gfm
The previous version of this was a custom implementation that I might have written based on Marked's implementation. That version continued to have issues, and I know regular expressions were likely insufficient for large parts of this, so I decided to start from scratch and attempt to port GitHub's cmark-gfm implementation.

Overall, it seems like that's gone pretty well. I had to learn to read some generated C code and understand GFM's extension system, but I'm confident that I got the implementation right. This initial version:
- Fixes MM-61148
- Adds support for nesting tables inside of block quotes matching GFM. The web app supports this but is less strict than GFM, and mobile doesn't currently support this.
- Adds support for nesting tables inside of lists, also matching GFM. Neither app currently supports this. This almost fixes MM-61148, but the specific example in that ticket has the table directly following a paragraph without a blank line in between which is something we don't support yet.